Implemented the support for proxy configuration from player
Implemented the reading of the RiseDisplayNetworkII.ini properties file
Passed the proxy configuration to the request module calls
Implemented handling of errors to connect through the proxy
Implemented handling of the errors to get the file or metadata from the host server. It will return 502 if not possible to connect, 504 if it cannot reach the server and 534 if it gets a 404 from the server.

Added tests and updated existent ones